+++ app/src/main/java/com/sopt/dive/util/CoroutineUtils.kt
@@ -1,0 +1,19 @@
+package com.sopt.dive.util
+
+import kotlinx.coroutines.TimeoutCancellationException
+import kotlinx.coroutines.currentCoroutineContext
+import kotlinx.coroutines.ensureActive
+import kotlin.coroutines.cancellation.CancellationException
+
+suspend fun <R> suspendRunCatching(block: suspend () -> R): Result<R> {
+    return try {
+        Result.success(block())
+    } catch (t: TimeoutCancellationException) {
+        Result.failure(t)
+    } catch (c: CancellationException) {
+        throw c
+    } catch (e: Throwable) {
+        currentCoroutineContext().ensureActive()
+        Result.failure(e)
+    }
+}
